refactor(api-compare): regenerate the narrow artifact inside api:calls
The narrow call-mismatch gate gated whatever output/call-mismatches.json
was on disk, so a sibling PR deleting a TS method surfaced as STALE
baseline entries on a branch that never touched it — the failure mode
RFC 0083 already removed from the wide gate.

Lift shouldRegenerate / regenerateArtifact out of the wide gate into
gate-regen.ts and have both gates import them, so the narrow gate reuses
the same opt-out contract (--no-regen, API_COMPARE_SKIP_WIDE_REGEN=1,
CI, and API_COMPARE_FORCE for --write) instead of inventing a second
one. regenerateArtifact now takes the compare scope as an argument.
