+++ scripts/api-compare/lint-call-mismatches-wide.test.ts
@@ -11,10 +11,6 @@ import {
   parseTop,
   relPathFor,
   renderReport,
-  shouldRegenerate,
-  regenerateArtifact,
-  NO_REGEN_FLAG,
-  REGEN_SKIP_ENV,
   unreviewedCount,
   writeSplitBaseline,
   renderStaleTags,
@@ -330,42 +326,6 @@ describe("parseTop", () => {
     expect(() => parseTop(["--top=x"], 20)).toThrow(/positive integer/);
     expect(() => parseTop(["--top=0"], 20)).toThrow(/positive integer/);
     expect(() => parseTop(["--top=2.5"], 20)).toThrow(/positive integer/);
-  });
-});
-
-describe("shouldRegenerate", () => {
-  it("regenerates on a plain local gating run", () => {
-    expect(shouldRegenerate([], {})).toBe(true);
-  });
-
-  it("opts out under CI, which runs compare.ts --wide-calls as its own step", () => {
-    expect(shouldRegenerate([], { CI: "true" })).toBe(false);
-  });
-
-  it("opts out on the explicit flag or env escape hatch", () => {
-    expect(shouldRegenerate([NO_REGEN_FLAG], {})).toBe(false);
-    expect(shouldRegenerate([], { [REGEN_SKIP_ENV]: "1" })).toBe(false);
-  });
-
-  it("leaves the read-only views alone", () => {
-    expect(shouldRegenerate(["--report"], {})).toBe(false);
-    expect(shouldRegenerate(["--unreviewed"], {})).toBe(false);
-  });
-
-  it("regenerates for a bare --write, so a reseed never baselines a stale artifact", () => {
-    expect(shouldRegenerate(["--write"], {})).toBe(true);
-  });
-
-  it("skips the --write regeneration only when the reseed script already forced one", () => {
-    expect(shouldRegenerate(["--write"], { API_COMPARE_FORCE: "1" })).toBe(false);
-  });
-});
-
-describe("regenerateArtifact", () => {
-  it("rejects with the failing command when the regeneration exits non-zero", async () => {
-    await expect(regenerateArtifact({ PATH: "" })).rejects.toThrow(
-      /pnpm api:compare --wide-calls|spawn/,
-    );
   });
 });
